create new utils.py file in pyrit/common
